Add java mcts with heuristic

+++ MKAgent/heuristics/Heuristic.java
@@ -4,5 +4,5 @@ import MKAgent.game.Board;
 import MKAgent.game.Side;
 
 public interface Heuristic {
-    int getScore(Board board, Side side);
+    double  getScore(Board board, Side side);
 }
