+++ MKAgent/heuristics/SeedsAwayFromStore.java
@@ -5,7 +5,7 @@ import MKAgent.game.Side;
 
 public class SeedsAwayFromStore implements Heuristic {
     @Override
-    public int getScore(Board board, Side side) {
+    public double getScore(Board board, Side side) {
         return getSeedsAwayFromStoreOnSide(board, side) - getSeedsAwayFromStoreOnSide(board, side.opposite());
     }
