+++ MKAgent/heuristics/SeedsDifference.java
@@ -5,7 +5,7 @@ import MKAgent.game.Side;
 
 public class SeedsDifference implements Heuristic {
     @Override
-    public int getScore(Board board, Side side) {
+    public double getScore(Board board, Side side) {
         return board.getSeedsInStore(side) - board.getSeedsInStore(side.opposite());
     }
 }
